Remove support for compund types completely
It turns out that these are confusing and implementation is error-prone.
Change the underlying Rust type to not allow a `Vec<ObjectType>` but just
`ObjectType` and basically relied on the compiler to tell me what to
fix.

Checked that the array type cases are unaffected - see tests in
`tests/data/arrays`.

Change-type: minor