Add basic fuzzer configuration.
Add standard `cargo-fuzz` configuration with seeds.
This just tests if given random input the whole stack does not panic.
If there is `Err` it is considered ok.
Please see `docs/MAINTAINER.md` for documentation on running.
I've ran this for half an hour and it did not find any crashes.

The plan would be to have this, and then as the next step (next PR) add
a fuzzing target that would use the new validator library, to check if
there are no `Err` returned if the input yaml is a valid jellyschema -
here I suspect it will find some corener cases.

Change-type: patch